Fix silent omission of images by autogen
When autogen moved to a new line, it would omit one image from the
resulting  template. This was clearly unintentional and a bug in the
jinja2 template.

The {% else %} statement in jinja is ambiguous because it defines the
default statement not only for "if" blocks but also "for" loops, which
is not how it works in actual python and most other languages. In this
case it could just have been merged with the if in the next line and
become an {% elif %}. The result would be functionally the same but it
would make the template more readable and the bug more obvious.

fixes #114